refactor: Move Settings to SettingsManager
