+++ include/Core/SettingsManager.hpp
