+++ src/Preferences/PreferencesPageTemplate.cpp
@@ -16,7 +16,7 @@
  */
 
 #include "Preferences/PreferencesPageTemplate.hpp"
-#include "Core/Settings.hpp"
+#include "Core/SettingsManager.hpp"
 #include "Util.hpp"
 #include <QDebug>
 #include <generated/SettingsInfo.hpp>
